gw: make SmartGatewayDividerEtx a 32 bit number
We need bigger dividers.

+++ src/gateway_costs.h
@@ -37,7 +37,7 @@
  * WexitU    = [1,           255] =  8 bits
  * WexitD    = [1,           255] =  8 bits
  * Wetx      = [1,           255] =  8 bits
- * Detx      = [1,           255] =  8 bits
+ * Detx      = [1, 4,294,967,295] = 32 bits
  * path_cost = [1, 4,294,967,295] = 32 bits
  *
  *                         1000 * 255   1000 * 255   255
@@ -64,7 +64,7 @@ extern "C" {
       uint8_t WexitU;
       uint8_t WexitD;
       uint8_t Wetx;
-      uint8_t Detx;
+      uint32_t Detx;
   };
 
   /**
